Construct the Rekor LogID in the reconciler instead of validating webhook.
