+++ pkg/reconciler/trustroot/trustroot.go
@@ -16,6 +16,11 @@ package trustroot
 
 import (
 	"context"
+	"crypto"
+	"crypto/ecdsa"
+	"crypto/sha256"
+	"crypto/x509"
+	"encoding/hex"
 	"errors"
 	"fmt"
 
@@ -23,6 +28,7 @@ import (
 	"github.com/sigstore/policy-controller/pkg/apis/policy/v1alpha1"
 	trustrootreconciler "github.com/sigstore/policy-controller/pkg/client/injection/reconciler/policy/v1alpha1/trustroot"
 	"github.com/sigstore/policy-controller/pkg/reconciler/trustroot/resources"
+	"github.com/sigstore/sigstore/pkg/cryptoutils"
 
 	corev1 "k8s.io/api/core/v1"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
@@ -68,6 +74,28 @@ func (r *Reconciler) ReconcileKind(ctx context.Context, trustroot *v1alpha1.Trus
 		logging.FromContext(ctx).Errorf("Failed to get Sigstore Keys: %v", err)
 		return err
 	}
+	// LogIDs for Rekor get created from the PublicKey, so we need to construct
+	// them before serializing.
+	for i, tlog := range sigstoreKeys.TLogs {
+		pk, err := cryptoutils.UnmarshalPEMToPublicKey(tlog.PublicKey)
+		if err != nil {
+			return fmt.Errorf("unmarshaling public key %d failed: %w", i, err)
+		}
+		// This needs to be ecdsa instead of crypto.PublicKey
+		// https://github.com/sigstore/cosign/issues/2540
+		pkecdsa, ok := pk.(*ecdsa.PublicKey)
+		if !ok {
+			return fmt.Errorf("public key %d is not ecdsa.PublicKey", i)
+		}
+		if tlog.LogID == "" {
+			logID, err := getLogID(tlog.LogID, pkecdsa)
+			if err != nil {
+				return fmt.Errorf("failed to construct LogID for rekor: %w", err)
+			}
+			sigstoreKeys.TLogs[i].LogID = logID
+		}
+	}
+
 	// See if the CM holding configs exists
 	existing, err := r.configmaplister.ConfigMaps(system.Namespace()).Get(config.SigstoreKeysConfigName)
 	if err != nil {
@@ -153,4 +181,21 @@ func (r *Reconciler) removeTrustRootEntry(ctx context.Context, cm *corev1.Config
 		return err
 	}
 	return nil
+}
+
+// LogID for Rekor apparently gets generated from the public key, so for future
+// proofing, we allow one to specify it in the TLog config for
+// TransparencyLogInstance but perhaps for Rekor we should never allow for it?
+//
+// getLogID generates a SHA256 hash of a DER-encoded public key.
+func getLogID(logID string, pub crypto.PublicKey) (string, error) {
+	if logID != "" {
+		return logID, nil
+	}
+	pubBytes, err := x509.MarshalPKIXPublicKey(pub)
+	if err != nil {
+		return "", err
+	}
+	digest := sha256.Sum256(pubBytes)
+	return hex.EncodeToString(digest[:]), nil
 }
