+++ pkg/reconciler/trustroot/trustroot_test.go
@@ -95,6 +95,9 @@ MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE7D2WvgqSzs9jpdJsOJ5Nl6xg8JXm
 Nmo7M3bN7+dQddw9Ibc2R3SV8tzBZw0rST8FKcn4apJepcKM4qUpYUeNfw==
 -----END PUBLIC KEY-----
 `
+	// This is the Rekor LogID constructed from above public key.
+	rekorLogID = "0bac0fddd0c15fbc46f8b1bf51c2b57676a9f262294fe13417d85602e73f392a"
+
 	tsaCertChain = `-----BEGIN CERTIFICATE-----
 MIIBzDCCAXKgAwIBAgIUfyGKDoFa7y6s/W1p1CiTmBRs1eAwCgYIKoZIzj0EAwIw
 MDEOMAwGA1UEChMFbG9jYWwxHjAcBgNVBAMTFVRlc3QgVFNBIEludGVybWVkaWF0
@@ -134,7 +137,7 @@ pSTJsAwzjW78CKQm7qol0uPmPPu6mNaw
 
 	// This is the marshalled entry from above keys/certs with fixed values
 	// (for ease of testing) for other parts.
-	marshalledEntry = `{"certificateAuthorities":[{"subject":{"organization":"fulcio-organization","commonName":"fulcio-common-name"},"uri":"https://fulcio.example.com","certChain":"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUZ3ekNDQTZ1Z0F3SUJBZ0lJSzd4YitycVk0Z0V3RFFZSktvWklodmNOQVFFTEJRQXdmakVNTUFvR0ExVUUKQmhNRFZWTkJNUk13RVFZRFZRUUlFd3BEWVd4cFptOXlibWxoTVJZd0ZBWURWUVFIRXcxVFlXNGdSbkpoYm1OcApjMk52TVJZd0ZBWURWUVFKRXcwMU5EZ2dUV0Z5YTJWMElGTjBNUTR3REFZRFZRUVJFd1UxTnpJM05ERVpNQmNHCkExVUVDaE1RVEdsdWRYZ2dSbTkxYm1SaGRHbHZiakFlRncweU1qRXlNRGd3TWpFM05URmFGdzB5TXpFeU1EZ3cKTWpFM05URmFNSDR4RERBS0JnTlZCQVlUQTFWVFFURVRNQkVHQTFVRUNCTUtRMkZzYVdadmNtNXBZVEVXTUJRRwpBMVVFQnhNTlUyRnVJRVp5WVc1amFYTmpiekVXTUJRR0ExVUVDUk1OTlRRNElFMWhjbXRsZENCVGRERU9NQXdHCkExVUVFUk1GTlRjeU56UXhHVEFYQmdOVkJBb1RFRXhwYm5WNElFWnZkVzVrWVhScGIyNHdnZ0lpTUEwR0NTcUcKU0liM0RRRUJBUVVBQTRJQ0R3QXdnZ0lLQW9JQ0FRQzE0MkVqbGcyUXhJd3BOamJhZVcvZnQ5c0gxVFhVNkNXZwpic3ZWcDc3dlJnY2tTbnBNM1JUQy9nd0V3Skh0WCtHT1RyUDlybzZuRkpOM0czaGNGbmFNSExLZEdyb2Y5aUh1Ci93L2xaTHdRelh6VlQrMFp5Wnh5dEhBV0dGQnZtWU00SjMzakg2RGo5UHZxT053dFNCU21aQlBjL0gvOEV2WXMKVXp4UFd1a2hPdG90U0gzVlhEcVo0amw5Nk1MZTArNWcyV2k3TXhSWDQ0WDFSaVBTMTRiYTFFUzUzOGJUaGhjUQo0U01qM3VoYmRzQ0lrY203ZUY0RVkzcEVYUXBYRUVHblpHZndZZ1FyKzZjVDA3WmQvV0RNME5YM0t4SDZxUms5CmdEalBuZmNNdUZiT1RiZkQvbnV2eDZGTlg2T1VyenJaU2dsa0x2Y1BJQlZPVzdMbjQxTEFiN2FYbWJXTEZFSm4KdUxvb1BwWVlyKzZOaG5GRE5HcHNCS0dLci9rdmJReURLS3N0M0NLajlvdFBTMTM2M25pNDFxbm9BN1lXU3F4dwp6NDE4NWRLS2MrWTd5dkpRc1JscjZxRzFzTkxPK2M3N2ZTUzVWWkltek5vekJjUmt1TEpGbFgrV0IwdXpnUVU1CnM0NUlaVytmSzkybmZ1OE1tS2p6SFIraWR5cjRPeWpTMFlTTjNHTWdjMFVQN0s2aFZwaExlZEFwRnB5a0JTRkcKVWdpUFp3clQrbUdTVmdtT1hxNW4xZFFUQ0QxNGxFaDJxdDMvcmZmOHpOYzBDTUFOV3liYU1HQkdRNGJoVlZYZQpSS1l4OXUyUFpqUHY1M3A3WWIvRENkcW5HRUR3L0hDQkRpQ3M0b1llNGRhRTM2eFVvanhEU20zRGFlTkc2OHo5ClJMN2dmVWpBeFFJREFRQUJvMFV3UXpBT0JnTlZIUThCQWY4RUJBTUNBUVl3RWdZRFZSMFRBUUgvQkFnd0JnRUIKL3dJQkFUQWRCZ05WSFE0RUZnUVVmK2xiTlgwV2g0aCtRMFNSdGhSSytLZkxqcUV3RFFZSktvWklodmNOQVFFTApCUUFEZ2dJQkFFaEpqYTBaU0t3WGNhT1hDWVJYVEUwNitKYnBlekk1TGV2QmhtYlJRSzc4OVJxMTBKZUFYYTdtCkVUb1JHbEdGTEgydURUMTFtc0ZLeU0zdjY3S2xFMVNZVmNxS21DbFlmSVZFWUgzTGEwdUkrOXJIWm5XZ2I0QmwKeTFCOHdibEtKemhZUUQ5WjRIL2dzK0JBc29SWDVWb0Z5SWdrTkJrMXAzZnRhVkNia1F2UzBPWXRZczVpdzRlSwpjSTcxL0lzVElUM1pwcGo5UjhJR3Nxd0xLZ3pmbnlOY0ZKZHorb2hjNlYyMlBqWk1FQkhDc0hQTzRhdjJMbFdLCjVZMWZsTCsyYnFUcWJtTy9iamZYMHc0WjFEdW9qUmNPWkY3U0g0TzNRdTJZNy82OWdIN0NwMG5pVkNtNXorUzUKMDExVjZQdk1qcm1pRSt4Vmt4TEhiWUVnb2NiRmhkNURjaU1DWHB2c3VEWm9qYUkzRlJFbUJxaUloS29raTNyYgp3dUVseWE3OGJNd2taMWtycDc2bldzbzQ3LzArNTFpby9XcmlBZHIwY2ptem9uaG83UnFJRTNEQzc3Q0VNa2FnClp2S1NtTDNzZmYrV05Tcm5QbHpuSzE5TkEyejRJbVc5TXN6cVByQ1RRR1AvL0JCdTdTYW16b2ZWTTlmNFBBSXIKRlRwblc2c0dkcEN6UDhFMFdVdTlCK3ZpS3J0Zk0vOXN4bkk5V2hmSlBkckVQMGlaVzN2aHd2Z1FiS2I1RDJPUwpVNG5yVm92NkJXci9CbmhRSzhJWG8xdHEzajhGQ1JJb2xlWE5oa3M0Z25rT2FEc1cyS3RWcXd0SzNpTzNCdlBiCkw1dzBnZExqd01Ma2VrNzJ5NjFYcXo1V3had05obDVZY21CS3VTdm1WU0h2QTY4QlZTYkIKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo="}],"tLogs":[{"baseURL":"https://rekor.example.com","hashAlgorithm":"sha-256","publicKey":"LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFN0QyV3ZncVN6czlqcGRKc09KNU5sNnhnOEpYbQpObW83TTNiTjcrZFFkZHc5SWJjMlIzU1Y4dHpCWncwclNUOEZLY240YXBKZXBjS000cVVwWVVlTmZ3PT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==","logID":"rekor-log-id"}],"ctLogs":[{"baseURL":"https://ctfe.example.com","hashAlgorithm":"sha-256","publicKey":"LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFSnZDSmk3MDdmdjV0TUoxVTJUVk1aK3VPNGRLRwphRWN2amxDa2dCQ0tYYnJrdW1aVjBtMGRTbEsxVjFneEVpeVE4eTZoazFNeEpOZTJBWnJaVXQ3YTR3PT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==","logID":"ctfe-log-id"}],"timestampAuthorities":[{"subject":{"organization":"tsa-organization","commonName":"tsa-common-name"},"uri":"https://tsa.example.com","certChain":"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUJ6RENDQVhLZ0F3SUJBZ0lVZnlHS0RvRmE3eTZzL1cxcDFDaVRtQlJzMWVBd0NnWUlLb1pJemowRUF3SXcKTURFT01Bd0dBMVVFQ2hNRmJHOWpZV3d4SGpBY0JnTlZCQU1URlZSbGMzUWdWRk5CSUVsdWRHVnliV1ZrYVdGMApaVEFlRncweU1qRXhNRGt5TURNeE16UmFGdzB6TVRFeE1Ea3lNRE0wTXpSYU1EQXhEakFNQmdOVkJBb1RCV3h2ClkyRnNNUjR3SEFZRFZRUURFeFZVWlhOMElGUlRRU0JVYVcxbGMzUmhiWEJwYm1jd1dUQVRCZ2NxaGtqT1BRSUIKQmdncWhrak9QUU1CQndOQ0FBUjNLY0R5OWp3QVJYMHJEdnlyK01HR2tHM24xT0EwTVU1K1ppRG1ndXNGeWs2VQo2Ym92S1dWTWZEOEo4TlRjSlpFMFJhWUpyOC9kRTlrZ2NJSVhsaE13bzJvd2FEQU9CZ05WSFE4QkFmOEVCQU1DCkI0QXdIUVlEVlIwT0JCWUVGSE5uNVIzYjNNdFVkU05yRk80OVE2WERWU25rTUI4R0ExVWRJd1FZTUJhQUZOTFMKNmdubzdPbSsrUXQ1eklhK0g5bzBIaVQyTUJZR0ExVWRKUUVCL3dRTU1Bb0dDQ3NHQVFVRkJ3TUlNQW9HQ0NxRwpTTTQ5QkFNQ0EwZ0FNRVVDSVFDRjBvbG9obnZkVXE2VDcvd1BrMTlaNWFRUC95eFJUakNXWXVobi9UQ3lIZ0lnCmF6VjNhaXI0R1JaYk45YmRZdGNRN0pVQUtxODlHT2h0RmZsNmtjb1ZVdlU9Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0KLS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUIwakNDQVhpZ0F3SUJBZ0lVWHBCbVlKRkZhR1czY0M4cDZiL0RIcjFpOElvd0NnWUlLb1pJemowRUF3SXcKS0RFT01Bd0dBMVVFQ2hNRmJHOWpZV3d4RmpBVUJnTlZCQU1URFZSbGMzUWdWRk5CSUZKdmIzUXdIaGNOTWpJeApNVEE1TWpBeU9UTTBXaGNOTXpJeE1UQTVNakF6TkRNMFdqQXdNUTR3REFZRFZRUUtFd1ZzYjJOaGJERWVNQndHCkExVUVBeE1WVkdWemRDQlVVMEVnU1c1MFpYSnRaV1JwWVhSbE1Ga3dFd1lIS29aSXpqMENBUVlJS29aSXpqMEQKQVFjRFFnQUVLRFBEUkl3RFMxWkN5bXViNnlhbkNHNW1hMHFEakxwTm9uRHZvb1NrUkhFZ1UwVE5pYmVKbjZNKwo1VzYwOGhDdzhud3V1Y01iWFE0MWtOZXVCZWV2eXFONE1IWXdEZ1lEVlIwUEFRSC9CQVFEQWdFR01CTUdBMVVkCkpRUU1NQW9HQ0NzR0FRVUZCd01JTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRk5MUzZnbm8KN09tKytRdDV6SWErSDlvMEhpVDJNQjhHQTFVZEl3UVlNQmFBRkIxbnZYcE5LN0F1UWxiSit5YTZuUFNxV2krVApNQW9HQ0NxR1NNNDlCQU1DQTBnQU1FVUNJR2l3cUNJMjl3N0M0VjhUbHRDc2k3MjhzNUR0a2xDUHlTREFTVVN1CmE1eTVBaUVBNDBJZmRsd2Y3VWo4cThOU0Q2WjRnLzBqczB0R05kTFNVSjFkby9Xb04wcz0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQotLS0tLUJFR0lOIENFUlRJRklDQVRFLS0tLS0KTUlJQmxEQ0NBVHFnQXdJQkFnSVVZWng5c1MxNEVuN1N1SERPSkpQNElQb3BNalV3Q2dZSUtvWkl6ajBFQXdJdwpLREVPTUF3R0ExVUVDaE1GYkc5allXd3hGakFVQmdOVkJBTVREVlJsYzNRZ1ZGTkJJRkp2YjNRd0hoY05Nakl4Ck1UQTVNakF5T1RNMFdoY05Nekl4TVRBNU1qQXpORE0wV2pBb01RNHdEQVlEVlFRS0V3VnNiMk5oYkRFV01CUUcKQTFVRUF4TU5WR1Z6ZENCVVUwRWdVbTl2ZERCWk1CTUdCeXFHU000OUFnRUdDQ3FHU000OUF3RUhBMElBQkFiQgpCMFNVOEc3NWhWSVVwaENoQTRuZk93TldQMzQ3VGpTY0lkc0VQcktWbisvWTFIbW1MSEpEalNmbit4aEVGb0VrCjdqcWdycW9uNDhpNHhibzd4QXVqUWpCQU1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUEJnTlZIUk1CQWY4RUJUQUQKQVFIL01CMEdBMVVkRGdRV0JCUWRaNzE2VFN1d0xrSld5ZnNtdXB6MHFsb3ZrekFLQmdncWhrak9QUVFEQWdOSQpBREJGQWlCZTVQNTZmb3FtRmNaQVZwRWVBT0ZackFsRWlxMDVDQ3BNTlloNUVqTHZtQUloQUtORjZ4SVY1dUZkCnBTVEpzQXd6alc3OENLUW03cW9sMHVQbVBQdTZtTmF3Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K"}]}`
+	marshalledEntry = `{"certificateAuthorities":[{"subject":{"organization":"fulcio-organization","commonName":"fulcio-common-name"},"uri":"https://fulcio.example.com","certChain":"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUZ3ekNDQTZ1Z0F3SUJBZ0lJSzd4YitycVk0Z0V3RFFZSktvWklodmNOQVFFTEJRQXdmakVNTUFvR0ExVUUKQmhNRFZWTkJNUk13RVFZRFZRUUlFd3BEWVd4cFptOXlibWxoTVJZd0ZBWURWUVFIRXcxVFlXNGdSbkpoYm1OcApjMk52TVJZd0ZBWURWUVFKRXcwMU5EZ2dUV0Z5YTJWMElGTjBNUTR3REFZRFZRUVJFd1UxTnpJM05ERVpNQmNHCkExVUVDaE1RVEdsdWRYZ2dSbTkxYm1SaGRHbHZiakFlRncweU1qRXlNRGd3TWpFM05URmFGdzB5TXpFeU1EZ3cKTWpFM05URmFNSDR4RERBS0JnTlZCQVlUQTFWVFFURVRNQkVHQTFVRUNCTUtRMkZzYVdadmNtNXBZVEVXTUJRRwpBMVVFQnhNTlUyRnVJRVp5WVc1amFYTmpiekVXTUJRR0ExVUVDUk1OTlRRNElFMWhjbXRsZENCVGRERU9NQXdHCkExVUVFUk1GTlRjeU56UXhHVEFYQmdOVkJBb1RFRXhwYm5WNElFWnZkVzVrWVhScGIyNHdnZ0lpTUEwR0NTcUcKU0liM0RRRUJBUVVBQTRJQ0R3QXdnZ0lLQW9JQ0FRQzE0MkVqbGcyUXhJd3BOamJhZVcvZnQ5c0gxVFhVNkNXZwpic3ZWcDc3dlJnY2tTbnBNM1JUQy9nd0V3Skh0WCtHT1RyUDlybzZuRkpOM0czaGNGbmFNSExLZEdyb2Y5aUh1Ci93L2xaTHdRelh6VlQrMFp5Wnh5dEhBV0dGQnZtWU00SjMzakg2RGo5UHZxT053dFNCU21aQlBjL0gvOEV2WXMKVXp4UFd1a2hPdG90U0gzVlhEcVo0amw5Nk1MZTArNWcyV2k3TXhSWDQ0WDFSaVBTMTRiYTFFUzUzOGJUaGhjUQo0U01qM3VoYmRzQ0lrY203ZUY0RVkzcEVYUXBYRUVHblpHZndZZ1FyKzZjVDA3WmQvV0RNME5YM0t4SDZxUms5CmdEalBuZmNNdUZiT1RiZkQvbnV2eDZGTlg2T1VyenJaU2dsa0x2Y1BJQlZPVzdMbjQxTEFiN2FYbWJXTEZFSm4KdUxvb1BwWVlyKzZOaG5GRE5HcHNCS0dLci9rdmJReURLS3N0M0NLajlvdFBTMTM2M25pNDFxbm9BN1lXU3F4dwp6NDE4NWRLS2MrWTd5dkpRc1JscjZxRzFzTkxPK2M3N2ZTUzVWWkltek5vekJjUmt1TEpGbFgrV0IwdXpnUVU1CnM0NUlaVytmSzkybmZ1OE1tS2p6SFIraWR5cjRPeWpTMFlTTjNHTWdjMFVQN0s2aFZwaExlZEFwRnB5a0JTRkcKVWdpUFp3clQrbUdTVmdtT1hxNW4xZFFUQ0QxNGxFaDJxdDMvcmZmOHpOYzBDTUFOV3liYU1HQkdRNGJoVlZYZQpSS1l4OXUyUFpqUHY1M3A3WWIvRENkcW5HRUR3L0hDQkRpQ3M0b1llNGRhRTM2eFVvanhEU20zRGFlTkc2OHo5ClJMN2dmVWpBeFFJREFRQUJvMFV3UXpBT0JnTlZIUThCQWY4RUJBTUNBUVl3RWdZRFZSMFRBUUgvQkFnd0JnRUIKL3dJQkFUQWRCZ05WSFE0RUZnUVVmK2xiTlgwV2g0aCtRMFNSdGhSSytLZkxqcUV3RFFZSktvWklodmNOQVFFTApCUUFEZ2dJQkFFaEpqYTBaU0t3WGNhT1hDWVJYVEUwNitKYnBlekk1TGV2QmhtYlJRSzc4OVJxMTBKZUFYYTdtCkVUb1JHbEdGTEgydURUMTFtc0ZLeU0zdjY3S2xFMVNZVmNxS21DbFlmSVZFWUgzTGEwdUkrOXJIWm5XZ2I0QmwKeTFCOHdibEtKemhZUUQ5WjRIL2dzK0JBc29SWDVWb0Z5SWdrTkJrMXAzZnRhVkNia1F2UzBPWXRZczVpdzRlSwpjSTcxL0lzVElUM1pwcGo5UjhJR3Nxd0xLZ3pmbnlOY0ZKZHorb2hjNlYyMlBqWk1FQkhDc0hQTzRhdjJMbFdLCjVZMWZsTCsyYnFUcWJtTy9iamZYMHc0WjFEdW9qUmNPWkY3U0g0TzNRdTJZNy82OWdIN0NwMG5pVkNtNXorUzUKMDExVjZQdk1qcm1pRSt4Vmt4TEhiWUVnb2NiRmhkNURjaU1DWHB2c3VEWm9qYUkzRlJFbUJxaUloS29raTNyYgp3dUVseWE3OGJNd2taMWtycDc2bldzbzQ3LzArNTFpby9XcmlBZHIwY2ptem9uaG83UnFJRTNEQzc3Q0VNa2FnClp2S1NtTDNzZmYrV05Tcm5QbHpuSzE5TkEyejRJbVc5TXN6cVByQ1RRR1AvL0JCdTdTYW16b2ZWTTlmNFBBSXIKRlRwblc2c0dkcEN6UDhFMFdVdTlCK3ZpS3J0Zk0vOXN4bkk5V2hmSlBkckVQMGlaVzN2aHd2Z1FiS2I1RDJPUwpVNG5yVm92NkJXci9CbmhRSzhJWG8xdHEzajhGQ1JJb2xlWE5oa3M0Z25rT2FEc1cyS3RWcXd0SzNpTzNCdlBiCkw1dzBnZExqd01Ma2VrNzJ5NjFYcXo1V3had05obDVZY21CS3VTdm1WU0h2QTY4QlZTYkIKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo="}],"tLogs":[{"baseURL":"https://rekor.example.com","hashAlgorithm":"sha-256","publicKey":"LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFN0QyV3ZncVN6czlqcGRKc09KNU5sNnhnOEpYbQpObW83TTNiTjcrZFFkZHc5SWJjMlIzU1Y4dHpCWncwclNUOEZLY240YXBKZXBjS000cVVwWVVlTmZ3PT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==","logID":"0bac0fddd0c15fbc46f8b1bf51c2b57676a9f262294fe13417d85602e73f392a"}],"ctLogs":[{"baseURL":"https://ctfe.example.com","hashAlgorithm":"sha-256","publicKey":"LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFSnZDSmk3MDdmdjV0TUoxVTJUVk1aK3VPNGRLRwphRWN2amxDa2dCQ0tYYnJrdW1aVjBtMGRTbEsxVjFneEVpeVE4eTZoazFNeEpOZTJBWnJaVXQ3YTR3PT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==","logID":"ctfe-log-id"}],"timestampAuthorities":[{"subject":{"organization":"tsa-organization","commonName":"tsa-common-name"},"uri":"https://tsa.example.com","certChain":"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUJ6RENDQVhLZ0F3SUJBZ0lVZnlHS0RvRmE3eTZzL1cxcDFDaVRtQlJzMWVBd0NnWUlLb1pJemowRUF3SXcKTURFT01Bd0dBMVVFQ2hNRmJHOWpZV3d4SGpBY0JnTlZCQU1URlZSbGMzUWdWRk5CSUVsdWRHVnliV1ZrYVdGMApaVEFlRncweU1qRXhNRGt5TURNeE16UmFGdzB6TVRFeE1Ea3lNRE0wTXpSYU1EQXhEakFNQmdOVkJBb1RCV3h2ClkyRnNNUjR3SEFZRFZRUURFeFZVWlhOMElGUlRRU0JVYVcxbGMzUmhiWEJwYm1jd1dUQVRCZ2NxaGtqT1BRSUIKQmdncWhrak9QUU1CQndOQ0FBUjNLY0R5OWp3QVJYMHJEdnlyK01HR2tHM24xT0EwTVU1K1ppRG1ndXNGeWs2VQo2Ym92S1dWTWZEOEo4TlRjSlpFMFJhWUpyOC9kRTlrZ2NJSVhsaE13bzJvd2FEQU9CZ05WSFE4QkFmOEVCQU1DCkI0QXdIUVlEVlIwT0JCWUVGSE5uNVIzYjNNdFVkU05yRk80OVE2WERWU25rTUI4R0ExVWRJd1FZTUJhQUZOTFMKNmdubzdPbSsrUXQ1eklhK0g5bzBIaVQyTUJZR0ExVWRKUUVCL3dRTU1Bb0dDQ3NHQVFVRkJ3TUlNQW9HQ0NxRwpTTTQ5QkFNQ0EwZ0FNRVVDSVFDRjBvbG9obnZkVXE2VDcvd1BrMTlaNWFRUC95eFJUakNXWXVobi9UQ3lIZ0lnCmF6VjNhaXI0R1JaYk45YmRZdGNRN0pVQUtxODlHT2h0RmZsNmtjb1ZVdlU9Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0KLS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUIwakNDQVhpZ0F3SUJBZ0lVWHBCbVlKRkZhR1czY0M4cDZiL0RIcjFpOElvd0NnWUlLb1pJemowRUF3SXcKS0RFT01Bd0dBMVVFQ2hNRmJHOWpZV3d4RmpBVUJnTlZCQU1URFZSbGMzUWdWRk5CSUZKdmIzUXdIaGNOTWpJeApNVEE1TWpBeU9UTTBXaGNOTXpJeE1UQTVNakF6TkRNMFdqQXdNUTR3REFZRFZRUUtFd1ZzYjJOaGJERWVNQndHCkExVUVBeE1WVkdWemRDQlVVMEVnU1c1MFpYSnRaV1JwWVhSbE1Ga3dFd1lIS29aSXpqMENBUVlJS29aSXpqMEQKQVFjRFFnQUVLRFBEUkl3RFMxWkN5bXViNnlhbkNHNW1hMHFEakxwTm9uRHZvb1NrUkhFZ1UwVE5pYmVKbjZNKwo1VzYwOGhDdzhud3V1Y01iWFE0MWtOZXVCZWV2eXFONE1IWXdEZ1lEVlIwUEFRSC9CQVFEQWdFR01CTUdBMVVkCkpRUU1NQW9HQ0NzR0FRVUZCd01JTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRk5MUzZnbm8KN09tKytRdDV6SWErSDlvMEhpVDJNQjhHQTFVZEl3UVlNQmFBRkIxbnZYcE5LN0F1UWxiSit5YTZuUFNxV2krVApNQW9HQ0NxR1NNNDlCQU1DQTBnQU1FVUNJR2l3cUNJMjl3N0M0VjhUbHRDc2k3MjhzNUR0a2xDUHlTREFTVVN1CmE1eTVBaUVBNDBJZmRsd2Y3VWo4cThOU0Q2WjRnLzBqczB0R05kTFNVSjFkby9Xb04wcz0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQotLS0tLUJFR0lOIENFUlRJRklDQVRFLS0tLS0KTUlJQmxEQ0NBVHFnQXdJQkFnSVVZWng5c1MxNEVuN1N1SERPSkpQNElQb3BNalV3Q2dZSUtvWkl6ajBFQXdJdwpLREVPTUF3R0ExVUVDaE1GYkc5allXd3hGakFVQmdOVkJBTVREVlJsYzNRZ1ZGTkJJRkp2YjNRd0hoY05Nakl4Ck1UQTVNakF5T1RNMFdoY05Nekl4TVRBNU1qQXpORE0wV2pBb01RNHdEQVlEVlFRS0V3VnNiMk5oYkRFV01CUUcKQTFVRUF4TU5WR1Z6ZENCVVUwRWdVbTl2ZERCWk1CTUdCeXFHU000OUFnRUdDQ3FHU000OUF3RUhBMElBQkFiQgpCMFNVOEc3NWhWSVVwaENoQTRuZk93TldQMzQ3VGpTY0lkc0VQcktWbisvWTFIbW1MSEpEalNmbit4aEVGb0VrCjdqcWdycW9uNDhpNHhibzd4QXVqUWpCQU1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUEJnTlZIUk1CQWY4RUJUQUQKQVFIL01CMEdBMVVkRGdRV0JCUWRaNzE2VFN1d0xrSld5ZnNtdXB6MHFsb3ZrekFLQmdncWhrak9QUVFEQWdOSQpBREJGQWlCZTVQNTZmb3FtRmNaQVZwRWVBT0ZackFsRWlxMDVDQ3BNTlloNUVqTHZtQUloQUtORjZ4SVY1dUZkCnBTVEpzQXd6alc3OENLUW03cW9sMHVQbVBQdTZtTmF3Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K"}]}`
 
 	replacePatchFmtString = `[{"op":"replace","path":"/data/%s","value":"%s"}]`
 	removePatchFmtString  = `[{"op":"remove","path":"/data/%s"}]`
@@ -286,6 +289,9 @@ func makeConfigMapWithSigstoreKeys() *corev1.ConfigMap {
 	source := NewTrustRoot(trName, WithSigstoreKeys(sigstoreKeys))
 	c := &config.SigstoreKeys{}
 	c.ConvertFrom(context.Background(), source.Spec.SigstoreKeys)
+	for i := range c.TLogs {
+		c.TLogs[i].LogID = rekorLogID
+	}
 	marshalled, err := resources.Marshal(c)
 	if err != nil {
 		panic("failed to marshal test SigstoreKeys")
